refactor(core): 주문 백그라운드 트리거를 infra로 이동
