+++ core/core-infra/src/main/java/com/ticket/core/infra/order/HoldReleaseOutboxScheduler.java
@@ -1,4 +1,9 @@
-package com.ticket.core.domain.order.command.release;
+package com.ticket.core.infra.order;
+
+import com.ticket.core.domain.order.command.release.HoldReleaseOutbox;
+import com.ticket.core.domain.order.command.release.HoldReleaseOutboxExecutor;
+import com.ticket.core.domain.order.command.release.HoldReleaseOutboxRepository;
+import com.ticket.core.domain.order.command.release.HoldReleaseOutboxStatus;
 
 import com.ticket.core.support.lock.DistributedLock;
 import lombok.RequiredArgsConstructor;
