+++ core/core-infra/src/main/java/com/ticket/core/infra/order/OrderExpirationScheduler.java
@@ -1,5 +1,6 @@
-package com.ticket.core.domain.order.command.expire;
+package com.ticket.core.infra.order;
 
+import com.ticket.core.domain.order.command.expire.ExpireOrderUseCase;
 import com.ticket.core.domain.order.model.Order;
 import com.ticket.core.domain.order.repository.OrderRepository;
 import com.ticket.core.domain.order.model.OrderState;
